Create 2.2 Basic Queue Operations
Play around with a queue. You will be given an integer N representing the number of elements to enqueue (add), an integer S, representing the number of elements to dequeue (remove) from the queue, and finally an integer X, an element that you should look for in the queue. If it is, print true on the console. If it's not printed the smallest element is currently present in the queue. If there are no elements in the sequence, print 0 on the console.