Create 2.8 Balanced Parenthesis
Given a sequence consisting of parentheses, determine whether the expression is balanced. A sequence of parentheses is balanced, if every open parenthesis can be paired uniquely with a closing parenthesis that occurs after the former. Also, the interval between them must be balanced. You will be given three types of parentheses: (, {, and [.