Create 51.5 Square With Maximum Sum
Create a program that reads a matrix from the console. Then find the biggest sum of the 2x2 submatrix and print it to the console.
On the first line, you will get matrix sizes in format rows, columns.
On the next rows lines, you will get elements for each column, separated with a comma and a space.
Print the biggest top-left square, which you find, and the sum of its elements.