Create 2.6 Jagged Array Manipulator
Create a program that populates, analyzes and manipulates the elements of a matrix with an unequal length of its rows.
First, you will receive an integer N equal to the number of rows in your matrix.
On the next N lines, you will receive sequences of integers, separated by a single space. Each sequence is a row in the matrix.
After populating the matrix, start analyzing it. If a row and the one below it have equal length, multiply each element in both of them by 2, otherwise - divide by 2.
Then, you will receive commands. There are three possible commands:
•	"Add {row} {column} {value}" - add {value} to the element at the given indexes, if they are valid.
•	"Subtract {row} {column} {value}" - subtract {value} from the element at the given indexes, if they are valid.
•	"End" - print the final state of the matrix (all elements separated by a single space) and stop the program.