Create 3.1 Blacksmith
First, you will be given a sequence representing steel. Afterward, you will be given another sequence representing carbon.
You should start from the first steel and try to mix it with the last carbon. If the sum of their values is equal to any of the swords in the table below you should forge the sword corresponding to the value and remove both the steel and the carbon. Otherwise, remove only the steel, increase the value of the carbon by 5 and insert it back into the collection. You need to stop forging when you have no more steel or carbon left.
Sword	Resources needed
Gladius	70
Shamshir	80
Katana	90
Sabre	110
Broadsword	150

Forge as many swords as possible.