Create 9.2 Warships
We get as input the size of the field in which our Warships are situated. The field is always a square. After that we will receive the coordinates for the attacks for each player in pairs of integers (row, column). If there is an attack on invalid coordinates - ignore the attack. The possible characters that may appear on the screen are:
•	* – a regular position on the field
•	< – ship of the first player.
•	> – ship of the second player
•	# – a sea mine that explodes when attacked
Each time when an attack hits a ship or a mine replace it with 'X'. Keep track of the count of ships remaining for each player. If a player hits a mine it explodes destroying all ships in adjacent fields (friend or foe). If a player destroys all enemy ships, the program stops and you have to print the following message: "Player {One/Two} has won the game! {totalCountShipsDestroyed} ships have been sunk in the battle."
If the list of attack commands ends before any of the player has won, you have to print the following message:     
"It's a draw! Player One has {countOfShips} left. Player Two has {countOfShips} left."