Create 10.1 Cooking
First, you will be given a sequence of integers, representing liquids. Afterwards, you will be given another sequence of integers representing ingredients.
You need to start from the first liquid and try to mix it with the last ingredient. If the sum of their values is equal to any of the items in the table below – cook the food corresponding to the value and remove both the liquid and the ingredient. Otherwise, remove only the liquid and increase the value of the ingredient by 3. You need to stop cooking when you have no more liquids or ingredients.
Food	Value needed
Bread	25
Cake	50
Pastry	75
Fruit Pie	100

In order to cook enough food for the bakery, you need one of each of the foods.