Create 1.4 Team
Create a Team class. Add to this team all of the people you have received. Those who are younger than 40 go to the first team, others go to the reserve team. At the end print the sizes of the first and the reserved team.
The class should have private fields for:
•	name: string
•	firstTeam: List<Person>
•	reserveTeam: List<Person>
The class should have constructors:
•	Team(string name)
The class should also have public properties for:
•	FirstTeam: List<Person> (read-only!)
•	ReserveTeam: List<Person> (read-only!)
And a method for adding players:
•	AddPlayer(Person person): void