Create 2.2 Animal Farm
You should be familiar with encapsulation already. For this problem, you’ll be working with the AnimalFarm project. It contains a class Chicken. Chicken contains several fields, a constructor, several properties, and methods. Your task is to encapsulate or hide anything unintended for viewing or modification from outside the class.