Create 2.6** Football Team Generator
A football Team has a variable number of players, a name, and a rating. A Player has a name and stats, which are the basis for his skill level. The stats a player has are endurance, sprint, dribble, passing, and shooting. Each stat can be an integer in the range [0..100]. The overall skill level of a player is calculated as the average of his stats. Only the name of a player and his stats should be visible to the entire outside world. Everything else should be hidden.
A Team should expose a name, a rating (calculated by the average skill level of all players in the team and rounded to the integer part only), and methods for adding and removing players.
Your task is to model the Team and the Player classes following the proper principles of Encapsulation. Expose only the properties that need to be visible and validate data appropriately.